Change trailing metadata delimiter
MarkdownAccessor accepts either "..." or "---" as the trailing delimiter
for metadata according to the Pandoc convention for metadata blocks.
However, the defacto standard is to use "---" for both delimiters.

This changes the trailing delimiter to "---" on output, while
still accepting both formats on input, to improve interoperability
with software that only accepts "---".

It also introduces the convention that the trailing delimiter
will be followed by a blank line.